Two primary parts of the update:
1) Continued improvements to the Contact CoCo Temp page by altering font-sizes, families, colors and positions on the page for a cleaner look.

2) Added a close-to-working reset password function to the site. This was done by adding a link on the login form page to resetting a user's password. On the new forgotPassword page, users can enter in their email and have an have an email sent to them with a link for resetting their password.

Internally, this is done by creating a PasswordResetToken that lasts for 24 hours. The token has a unique, generated UUID code that is used to authorize a user to reset their password. The various new Controllers and Services, namely the ResetPasswordController and the IUserToolsService interface / UserToolsService class are used to construct and send an email to the given email address, but not before checking if the email address already exists in our system.

The user never enters in the unique UUID code, but rather clicks on a unique link to the CoCoTemp site in the email. This link will check the token associated with the email, and if it is found to be valid, the user is temporarily given the privilege to change their password. The user is then directed to a new page, which is one of the final pieces of this functionality that still needs to be added, where they can enter in their new password.

Alongside the actual change password page, the functionality to change a user's password still needs to be added, and the various new classes and interfaces still need a bit of cleaning up in order to function properly.